+++ src/main/java/com/quicklybly/vacationcalendar/dto/CalculateResponse.java
@@ -3,5 +3,7 @@ package com.quicklybly.vacationcalendar.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
-public record CalculateResponse(@NotNull @PositiveOrZero Long vacationPay) {
+import java.math.BigDecimal;
+
+public record CalculateResponse(@NotNull @PositiveOrZero BigDecimal vacationPay) {
 }
